backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "fastapi[standard]>=0.121.3",
    "psycopg[binary]>=3.2.13",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.6",
    "sqlalchemy[asyncio]>=2.0.44",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
